feat: rename evm_bridge to mock_evm_bridge (#101)
